requirements.txt
chardet==5.2.0
pillow==11.1.0
reportlab==4.3.1
